Check for new posts more agressively
I do the crossword sometimes and notice the reddit post exists, but has
not yet been scraped yet. In this commit, I add additional times for the
scraper GH action to run, so we can publish posts sooner to users on the
site.

Note that the time the reddit posts are made isn't an exact science. It's
a non-contractual, best-effort routine done by the reddit mods. And even
then, they probably use some moderation software that cannot precisely
schedule posts anyway. So we have to flex to that.

Fortunately, this GH action will only add a new commit if there are new
posts found, so we won't dirty our history any more than we have to.